Change Latin Square version to work with PsiTurk.
Latin Square version used to be random. Now it's grabbed from the
webpage. PsiTurk will supply it as a variable called condition. It also
allows the use of another variable called counterbalance but I haven't
created a way of using both. It is no longer possible to have different
group sizes in different latin squared blocks.